Use CDI directly - byte readable
CDI is byte readable (but word writable) so we can use the address of
CDI directly as input when generating the key pair. This way we don't
have to keep a local copy on the stack at all.

Note that QEMU will warn for writes of the wrong size since
crypto_ed25519_key_pair() calls crypto_wipe() on the seed after using
it.