Refactor: Introduce protocol state machine
Introduce an explicit state machine controlling the protocol handling.

States:

- STATE_STARTED: Allows some info gathering commands, name/version,
  get public key, set size of message, or load a fix-sized pre-hashed
  message in a single command.
- STATE_LOADING: Allows loading a message (up to 4 kByte) to sign.
- STATE_SIGNING: Allows asking for the signature, which triggers the
  signing operation, and return of the signature.
- STATE_FAILED: parse error or state failure. Never returns.

We make sure to often move to STATE_FAILED on anything suspicious
which then halts the CPU until powercycled. For instance we only allow
firmware probes in STATE_STARTED and entering STATE_FAILED on any
probes in another state and any parse errors or unexpected lengths in
any states.

See docs/implementation-notes.md for more details.

One might argue that we should sign and return the signature
unconditionally when the last block of the message has been received,
but we'll have to wait to do that change since that changes the
application protocol. This would break any client apps depending on
the signer without warning.

Other refactoring that was done simultaneously with the above work
include:

- Use memory writing functions with destination size whenever
  possible.
- Less use of magic numbers.
- Clean up code and simplify names of for instance protocol constants.

This work was done in tandem to updates in tkey-libs and needs modern
tkey-libs.

App version is now: 3.