Warning: Replace PH commands with firmware hash commands
Instead of supporting pre-hash Ed25519 signatures, we replace these
commands with an ability to get the device app to do a SHA512 hash
over specified sizes of the firmware, the exact same commands that the
verisigner app in tkey-verification does. This way we don't have to
maintain two almost identical device apps.

Since these commands just happened to have the exact same constants as
the command and response verisigner used for firmware hashing we
override them with this implementation.

Pre-hash signing has never been officially supported in any released
tkeysign Go package and a sampling of users showed no one else seems
to use it either. Furthermore, Ed25519ph seems to be frowned upon by
cryptographers. We prefer not to support it.

We realize that since we support very small messages to be signed and
have low bandwidth people will probably ask for a signature of a
digest anyway, but it seems important to keep the signer pretty
simple, and there are ways to mitigate an exploit of a hash collision.
We leave that up to the caller.